+++ chag/__init__.py
@@ -1,5 +1,3 @@
-# -*- coding: utf-8 -*-
-
 __author__ = 'Michael Dowling'
 __email__ = 'mtdowling@gmail.com'
 __version__ = '0.3.0'