+++ chag/main.py
@@ -1,20 +1,31 @@
 import json as j
 import subprocess
 import tempfile
-import time
+import os
 
 import click
-from chag.changelog import get_tag, get_tags
+import changelog
+import chag.git as git
 
 
-def git_is_clean():
-    try:
-        cmd = 'git diff --stat --exit-code'.split(' ')
-        subprocess.check_call(cmd)
-        cmd.append('--cached')
-        subprocess.check_call(cmd)
-    except subprocess.CalledProcessError as e:
-        raise click.ClickException('Your repository is not clean')
+def _load_changelog(f):
+    """Loads the provided file or attempts to find the changelog file"""
+    if f:
+        return f
+    for check in ['CHANGELOG', 'CHANGELOG.md', 'CHANGELOG.rst']:
+        check = os.getcwd() + '/' + check
+        if os.path.isfile(check):
+            return open(check, 'r+')
+    raise click.ClickException('Changelog file not provided and not found')
+
+
+def _get_message(m):
+    """Gets a message from a variable or EDITOR if not set"""
+    if not m:
+        with tempfile.NamedTemporaryFile(suffix=".tmp") as tmp:
+            subprocess.call([os.environ.get('EDITOR', 'vim'), tmp.name])
+            m = open(tmp.name, 'r').read()[:-1]
+    return m
 
 
 @click.group()
@@ -25,32 +36,36 @@ def main(ctx):
 
 @main.command()
 @click.option('--border', default='-', help='Repeated border character')
-@click.argument('file', type=click.File('rb'))
-@click.argument('tag')
-def contents(file, tag, border):
-    """Returns the changelog contents for the provided TAG from the FILE.
-    Pass 'latest' as the TAG to parse the first found tag.
+@click.option('-f', type=click.File('r'), help='Path to changelog')
+@click.option('-t', help='Tag to retrieve', default='latest')
+def contents(t, f, border):
+    """Returns the changelog contents for the provided -t tag, or the latest
+    tag if no -t option is provided.
 
     \b
     Examples:
-      chag contents CHANGELOG latest
-      chag contents CHANGELOG 0.1.0
-      chag contents --border='=' CHANGELOG 0.1.0
+      chag contents
+      chag contents -t 0.1.0
+      chag contents -f CHANGELOG -t latest
+      chag contents --border='='
     """
-    click.echo(get_tag(file, border, tag)['contents'])
+    f = _load_changelog(f)
+    click.echo(changelog.get_tag(f, border, t)['contents'])
 
 
 @main.command()
 @click.option('--border', default='-', help='Repeated border character')
 @click.option('--json/--no-json', default=False,
               help='Determines if JSON or just the version is returned.')
-@click.argument('file', type=click.File('rb'))
-@click.argument('tag')
-def get(file, tag, border, json):
-    """Prints out a specific tag number or JSON data for a specific TAG found
-    in FILE if the --json option is passed. Pass 'latest' as the TAG to parse
-    the first found tag entry in the changelog. The JSON output contains the
-    following keys:
+@click.option('-f', type=click.File('r'), help='Path to changelog')
+@click.option('-t', help='Tag to retrieve', default='latest')
+def get(t, f, border, json):
+    """Prints out a specific tag number or JSON data for a specific tag. Pass
+    'latest' as the -t option to parse the first found tag entry in the
+    changelog.
+
+    This command outputs JSON data when the --json option is provided. The
+    JSON output contains the following keys:
 
     \b
     - line_number: Line that the tag heading is found
@@ -58,112 +73,118 @@ def get(file, tag, border, json):
     - tag: Just the tag from the heading (e.g., '0.1.0')
     - contents: The text content of the changelog entry
 
-    \b
     Examples:
-      chag get CHANGELOG latest
-      chag get --border='=' CHANGELOG latest
-      chag get --json CHANGELOG 0.0.1
+      chag get -f CHANGELOG latest
+      chag get --border='=' latest
+      chag get --json 0.0.1
     """
+    f = _load_changelog(f)
     if json:
-        click.echo(j.dumps(get_tag(file, border, tag)))
+        click.echo(j.dumps(changelog.get_tag(f, border, t)))
     else:
-        click.echo(get_tag(file, border, tag)['tag'])
+        click.echo(changelog.get_tag(f, border, t)['tag'])
 
 
 @main.command()
+@click.option('-f', type=click.File('r'), help='Path to changelog')
 @click.option('--border', default='-', help='Repeated border character')
-@click.argument('file', type=click.File('rb'))
-def list(file, border):
-    """Lists the tag versions available in a changelog file
+def list(f, border):
+    """Lists the tag versions available in a changelog file.
 
     \b
     Examples:
-      chag list CHANGELOG
-      chag list --border='=' CHANGELOG
+      chag list -f CHANGELOG
+      chag list --border='='
     """
-    for tag in get_tags(file, border):
+    for tag in changelog.get_tags(_load_changelog(f), border):
         print(tag['tag'])
 
 
 @main.command()
 @click.option('--border', default='-', help='Repeated border character')
-@click.argument('file', type=click.File('rb'))
-@click.argument('heading')
-def update(file, heading, border):
-    """Replaces the latest changelog entry in FILE with the provided
-    HEADING (e.g., '1.0.1 ()'). If the last two characters of HEADING
+@click.option('-f', type=click.File('r'), help='Path to changelog')
+@click.option('-m', help='Heading to apply to the changelog')
+def update(m, f, border):
+    """Replaces the latest changelog entry with the provided -m message. If no
+    -m option is provided, an editor will open where you can enter the
+    changelog heading. The editor that is opened is based on the $EDITOR
+    environment variable. If the last two characters of the provided heading
     are '()', they will be converted to today's date formatted as
     '(YYYY-MM-DD)'.
 
     \b
     Examples:
-      chag update CHANGELOG.rst '1.0.1'
-      chag update CHANGELOG.rst '1.0.1 ()'
-      chag update --border='=' CHANGELOG.rst '1.0.1 ()'
+      chag update -f CHANGELOG.rst -m '1.0.1'
+      chag update -m '1.0.1 ()'
+      chag update --border='='
     """
-    # Replace the date if needed
-    if heading[-2::] == '()':
-        replacement = '(' + time.strftime('%Y-%m-%d') + ')'
-        heading = heading.replace('()', replacement)
-    found = get_tag(file, border, 'latest')
-    file.seek(0)
-    lines = [l.decode('utf-8') for l in file.readlines()]
-    lines[found['line_number']] = heading + "\n"
-    lines[found['line_number'] + 1] = (border * len(heading)) + "\n"
-    with open(file.name, 'w') as f:
-        f.write(''.join(lines))
+    heading = _get_message(m)
+    heading = changelog.update(_load_changelog(f), border, heading)
     click.echo('Updated first changelog entry to %s' % heading, err=True)
+
+
+@main.command()
+@click.option('--border', default='-', help='Repeated border character')
+@click.option('--github', help='Pass the "user/repo" of your GitHub project')
+@click.option('-f', type=click.File('r'), help='Path to changelog')
+@click.option('-m', help='Changelog entry to append')
+def append(f, border, github, m):
+    """Appends an entry to the first changelog entry. Pass the entry in the
+    -m option. If no -m option is provided, an editor will open where you can
+    enter the changelog message to append. The editor that is opened is based
+    on the $EDITOR environment variable.
+
+    \b
+    Examples:
+        chag append -m '* Updated this file'
+        chag append --github -m '* Updated this file'
+        chag append --github --border='=' -m '* Updated this file'
+        chag append
+    """
+    m = _get_message(m)
+    if github:
+        m = git.github_markdown(github, m)
+    changelog.append(_load_changelog(f), border, m)
+    click.echo('Appended to the first changelog entry:', err=True)
+    click.echo(m, err=True)
+
 
 @main.command()
 @click.option('--border', default='-', help='Repeated border character')
 @click.option('--v-prefix/--no-v-prefix', default=False)
 @click.option('--sign/--no-sign', default=False)
 @click.option('--force/--no-force', default=False)
-@click.argument('file', type=click.File('rb'))
-def tag(file, border, v_prefix, sign, force):
-    """Creates an annotated git tag from the latest changelog entry in the
-    provided FILE. The tag uses the version number found in the changelog
-    entry's header line and the contents of the changelog entry as the tag
-    annotation.
+@click.option('-f', type=click.File('r'), help='Path to changelog')
+def tag(f, border, v_prefix, sign, force):
+    """Creates an annotated git tag from the latest changelog entry. The tag
+    uses the version number found in the changelog entry's header line and the
+    contents of the changelog entry as the tag annotation.
 
     \b
     Examples:
-      chag tag /path/to/CHANGELOG.md
-      chag tag --v-prefix CHANGELOG.rst
-      chag tag --sign CHANGELOG.rst
-      chag tag --force CHANGELOG.md
-      chag tag --debug CHANGELOG.md
+      chag tag -f /path/to/CHANGELOG.md
+      chag tag --v-prefix
+      chag tag --sign
+      chag tag --force
+      chag tag --debug
     """
-    found = get_tag(file, border, 'latest')
+    f = _load_changelog(f)
+    found = changelog.get_tag(f, border, 'latest')
     if found['tag'] == 'Next':
         raise click.ClickException('Not tagging a "Next Release" entry!')
     click.echo('Ensuring git repository is clean with git diff', err=True)
-    git_is_clean()
-    # Use the appropriate tag name based on the --v-prefix option
+    git.is_clean()
     if v_prefix:
         tag_name = 'v' + found['tag']
     else:
         tag_name = found['tag']
-    # Build up the git tag argument list
-    args = ['git', 'tag', '-a', '-F', '-']
-    if force:
-        args.append('--force')
-    if sign:
-        args.append('--sign')
-    args.append(tag_name)
-
-    click.echo('Creating a git tag using: ' + ' '.join(args), err=True)
     click.echo("Using the following annotation:", err=True)
     click.echo('  ' + found['contents'].replace("\n", "\n    "), err=True)
-
-    with tempfile.TemporaryFile() as tmp:
-        tmp.write(found['contents'])
-        tmp.seek(0)
-        try:
-            subprocess.check_call(args, stdin=tmp)
-            click.echo('[SUCCESS] Tagged %s' % tag_name, err=True)
-        except subprocess.CalledProcessError as e:
-            raise click.ClickException('[FAILED] %s' % str(e))
+    try:
+        git.get_tag_args(tag_name, force, sign, found['contents'])
+        click.echo('[SUCCESS] Tagged %s' % tag_name, err=True)
+    except Exception as e:
+        raise click.ClickException('[FAILED] %s' % str(e))
 
 
 if __name__ == '__main__':